rather than creating new kpis updated the existing ones
