+++ src/main/java/com/publicissapient/kpidashboard/apis/jira/scrum/service/CycleTimeServiceImpl.java
@@ -1,6 +1,5 @@
 package com.publicissapient.kpidashboard.apis.jira.scrum.service;
 
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -14,27 +13,29 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import com.publicissapient.kpidashboard.apis.enums.KPISource;
+import com.publicissapient.kpidashboard.apis.jira.service.JiraKPIService;
+import com.publicissapient.kpidashboard.apis.model.KPIExcelData;
+import com.publicissapient.kpidashboard.apis.model.TreeAggregatorDetail;
+import com.publicissapient.kpidashboard.apis.util.KPIExcelUtility;
+import com.publicissapient.kpidashboard.common.model.application.DataCountGroup;
+import com.publicissapient.kpidashboard.common.model.application.DataValue;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.ObjectUtils;
+import org.apache.commons.lang3.tuple.Pair;
 import org.bson.types.ObjectId;
 import org.joda.time.DateTime;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import com.publicissapient.kpidashboard.apis.appsetting.service.ConfigHelperService;
-import com.publicissapient.kpidashboard.apis.config.CustomApiConfig;
 import com.publicissapient.kpidashboard.apis.constant.Constant;
-import com.publicissapient.kpidashboard.apis.enums.JiraFeature;
 import com.publicissapient.kpidashboard.apis.enums.JiraFeatureHistory;
 import com.publicissapient.kpidashboard.apis.enums.KPICode;
 import com.publicissapient.kpidashboard.apis.enums.KPIExcelColumn;
 import com.publicissapient.kpidashboard.apis.errors.ApplicationException;
-import com.publicissapient.kpidashboard.apis.jira.service.backlogdashboard.JiraBacklogKPIService;
-import com.publicissapient.kpidashboard.apis.model.IterationKpiData;
 import com.publicissapient.kpidashboard.apis.model.IterationKpiFilters;
 import com.publicissapient.kpidashboard.apis.model.IterationKpiFiltersOptions;
-import com.publicissapient.kpidashboard.apis.model.IterationKpiModalValue;
-import com.publicissapient.kpidashboard.apis.model.IterationKpiValue;
 import com.publicissapient.kpidashboard.apis.model.KpiElement;
 import com.publicissapient.kpidashboard.apis.model.KpiRequest;
 import com.publicissapient.kpidashboard.apis.model.Node;
@@ -53,12 +54,13 @@ import com.publicissapient.kpidashboard.common.util.DateUtil;
 
 import lombok.extern.slf4j.Slf4j;
 
+import static com.publicissapient.kpidashboard.common.constant.CommonConstant.HIERARCHY_LEVEL_ID_PROJECT;
+
 @Component
 @Slf4j
-public class CycleTimeServiceImpl extends JiraBacklogKPIService<Long, List<Object>> {
+public class CycleTimeServiceImpl extends JiraKPIService<Double, List<Object>, Map<String, Object>> {
 	public static final String DAYS = "d";
 	public static final String ISSUES = "issues";
-	private static final String STORY_HISTORY_DATA = "storyHistoryData";
 	private static final String INTAKE_TO_DOR = "Intake - DOR";
 	private static final String DOR_TO_DOD = "DOR - DOD";
 	private static final String DOD_TO_LIVE = "DOD - Live";
@@ -68,101 +70,150 @@ public class CycleTimeServiceImpl extends JiraBacklogKPIService<Long, List<Objec
 	private static final String LEAD_TIME_KPI = "LEAD TIME";
 	private static final String SEARCH_BY_ISSUE_TYPE = "Issue Type";
 	private static final String SEARCH_BY_DURATION = "Duration";
+	private static final Set<String> durationFilter = new LinkedHashSet<>(
+			Arrays.asList("Past Week", "Past 2 Weeks", "Past Month", "Past 3 Months", "Past 6 Months"));
 	@Autowired
 	private JiraIssueCustomHistoryRepository jiraIssueCustomHistoryRepository;
 
 	@Autowired
 	private ConfigHelperService configHelperService;
 
-	@Autowired
-	private CustomApiConfig customApiConfig;
-
 	@Override
-	public KpiElement getKpiData(KpiRequest kpiRequest, KpiElement kpiElement, Node projectNode)
-			throws ApplicationException {
+	public KpiElement getKpiData(KpiRequest kpiRequest, KpiElement kpiElement,
+			TreeAggregatorDetail treeAggregatorDetail) throws ApplicationException {
 
 		log.info("LEAD-TIME -> requestTrackerId[{}]", kpiRequest.getRequestTrackerId());
-		DataCount dataCount = new DataCount();
-		projectWiseLeafNodeValue(projectNode, kpiElement, kpiRequest, dataCount);
-		log.debug("[LEAD-TIME-LEAF-NODE-VALUE][{}]. Values of leaf node after KPI calculation {}",
-				kpiRequest.getRequestTrackerId(), projectNode);
+		List<Node> projectList = treeAggregatorDetail.getMapOfListOfProjectNodes().get(HIERARCHY_LEVEL_ID_PROJECT);
+		Map<String, Node> mapTmp = treeAggregatorDetail.getMapTmp();
+
+		// in case if only projects or sprint filters are applied
+		projectWiseLeafNodeValue(kpiElement, mapTmp, projectList, kpiRequest);
+
+		Map<Pair<String, String>, Node> nodeWiseKPIValue = new HashMap<>();
+		calculateAggregatedMultipleValueGroup(treeAggregatorDetail.getRoot(), nodeWiseKPIValue, KPICode.CYCLE_TIME);
+		Map<String, List<DataCount>> trendValuesMap = getTrendValuesMap(kpiRequest, kpiElement, nodeWiseKPIValue,
+				KPICode.CYCLE_TIME);
+
+		Map<String, Map<String, List<DataCount>>> priorityTypeProjectWiseDc = new LinkedHashMap<>();
+		trendValuesMap.forEach((priority, dataCounts) -> {
+			Map<String, List<DataCount>> projectWiseDc = dataCounts.stream()
+					.collect(Collectors.groupingBy(DataCount::getData));
+			priorityTypeProjectWiseDc.put(priority, projectWiseDc);
+		});
+
+		List<DataCountGroup> dataCountGroups = new ArrayList<>();
+		priorityTypeProjectWiseDc.forEach((filter, projectWiseDc) -> {
+			DataCountGroup dataCountGroup = new DataCountGroup();
+			List<DataCount> dataList = new ArrayList<>();
+			projectWiseDc.forEach((key, value) -> dataList.addAll(value));
+			// split for filters
+			String[] issueFilter = filter.split("#");
+			dataCountGroup.setFilter1(issueFilter[0]);
+			dataCountGroup.setFilter2(issueFilter[1]);
+			dataCountGroup.setValue(dataList);
+			dataCountGroups.add(dataCountGroup);
+		});
+		kpiElement.setTrendValueList(dataCountGroups);
+		kpiElement.setNodeWiseKPIValue(nodeWiseKPIValue);
+
 		return kpiElement;
 	}
 
 	@SuppressWarnings("unchecked")
-	private void projectWiseLeafNodeValue(Node leafNode, KpiElement kpiElement, KpiRequest kpiRequest,
-			DataCount dataCount) {
+	private void projectWiseLeafNodeValue(KpiElement kpiElement, Map<String, Node> mapTmp, List<Node> leafNodeList,
+			KpiRequest kpiRequest) {
 		KpiElement leadTimeReq = kpiRequest.getKpiList().stream().filter(k -> k.getKpiId().equalsIgnoreCase("kpi171"))
 				.findFirst().orElse(new KpiElement());
 
 		LinkedHashMap<String, Object> filterDuration = (LinkedHashMap<String, Object>) leadTimeReq.getFilterDuration();
 		int value = 2; // Default value for 'value'
-		String duration = CommonConstant.WEEK; // Default value for 'duration'
+		String duration; // Default value for 'duration'
 		String startDate = null;
 		String endDate = DateUtil.getTodayDate().toString();
 
 		if (filterDuration != null) {
 			value = (int) filterDuration.getOrDefault("value", 6);
 			duration = (String) filterDuration.getOrDefault("duration", CommonConstant.MONTH);
-		}
-		if (duration.equalsIgnoreCase(CommonConstant.WEEK)) {
+		} else {
+            duration = CommonConstant.WEEK;
+        }
+        if (duration.equalsIgnoreCase(CommonConstant.WEEK)) {
 			startDate = DateUtil.getTodayDate().minusWeeks(value).toString();
 		} else if (duration.equalsIgnoreCase(CommonConstant.MONTH)) {
 			startDate = DateUtil.getTodayDate().minusMonths(value).toString();
 		}
-		Map<String, Object> resultMap = fetchKPIDataFromDb(leafNode, startDate, endDate, kpiRequest);
+		Map<String, List<String>> durationLabels = new HashMap<>();
+		durationLabels.put(CommonConstant.WEEK, Arrays.asList("Past Week", "Past 2 Weeks"));
+		durationLabels.put(CommonConstant.MONTH, Arrays.asList("Past Month", "Past 3 Months", "Past 6 Months"));
 
-		List<JiraIssueCustomHistory> ticketList = (List<JiraIssueCustomHistory>) resultMap.get(STORY_HISTORY_DATA);
+		String label = durationLabels.getOrDefault(duration, Collections.emptyList())
+				.stream()
+				.skip(value - 1L)
+				.findFirst()
+				.orElse("Unknown");
+		Map<String, Object> resultMap = fetchKPIDataFromDb(leafNodeList, startDate, endDate, kpiRequest);
+		List<CycleTimeValidationData> cycleTimeList = new ArrayList<>();
+		List<KPIExcelData> excelData = new ArrayList<>();
+		leafNodeList.forEach(leafNode -> {
+			List<JiraIssueCustomHistory> issueCustomHistoryList = (List<JiraIssueCustomHistory>) resultMap
+					.get(leafNode.getProjectFilter().getBasicProjectConfigId().toString());
 
-		Map<String, List<JiraIssueCustomHistory>> projectWiseJiraIssue = ticketList.stream()
-				.collect(Collectors.groupingBy(JiraIssueCustomHistory::getBasicProjectConfigId));
-
-		kpiWithFilter(projectWiseJiraIssue, leafNode, kpiElement, dataCount);
+			FieldMapping fieldMapping = configHelperService.getFieldMappingMap()
+					.get(leafNode.getProjectFilter().getBasicProjectConfigId());
+			Map<String, List<DataValue>> cycleTimeDataCount = getCycleTimeDataCount(issueCustomHistoryList, fieldMapping, cycleTimeList, kpiElement);
+			Map<String, List<DataCount>> dataCountMap = getDataCountObject(leafNode.getProjectFilter().getName(), cycleTimeDataCount, label);
+			mapTmp.get(leafNode.getId()).setValue(dataCountMap);
+		});
+		populateExcelDataObject(kpiRequest.getRequestTrackerId(), cycleTimeList, excelData);
+		kpiElement.setExcelData(excelData);
+		kpiElement.setModalHeads(KPIExcelColumn.CYCLE_TIME.getColumns());
 	}
 
 	@Override
-	public Map<String, Object> fetchKPIDataFromDb(Node leafNode, String startDate, String endDate,
+	public Double calculateKPIMetrics(Map<String, Object> stringObjectMap) {
+		return 0.0;
+	}
+
+	@Override
+	public Map<String, Object> fetchKPIDataFromDb(List<Node> leafNodeList, String startDate, String endDate,
 			KpiRequest kpiRequest) {
 		Map<String, Object> resultListMap = new HashMap<>();
-		Map<String, List<String>> mapOfFilters = new LinkedHashMap<>();
-		Map<String, Map<String, Object>> uniqueProjectMap = new HashMap<>();
+		leafNodeList.forEach(leafNode -> {
+			Map<String, Map<String, Object>> uniqueProjectMap = new HashMap<>();
 
-		List<String> basicProjectConfigIds = new ArrayList<>();
-		ObjectId basicProjectConfigId = leafNode.getProjectFilter().getBasicProjectConfigId();
-		Map<String, Object> mapOfProjectFilters = new LinkedHashMap<>();
+			ObjectId basicProjectConfigId = leafNode.getProjectFilter().getBasicProjectConfigId();
+			Map<String, Object> mapOfProjectFilters = new LinkedHashMap<>();
 
-		FieldMapping fieldMapping = configHelperService.getFieldMappingMap().get(basicProjectConfigId);
+			FieldMapping fieldMapping = configHelperService.getFieldMappingMap().get(basicProjectConfigId);
 
-		basicProjectConfigIds.add(basicProjectConfigId.toString());
+			if (Optional.ofNullable(fieldMapping.getJiraIssueTypeKPI171()).isPresent()) {
 
-		if (Optional.ofNullable(fieldMapping.getJiraIssueTypeKPI171()).isPresent()) {
+				KpiDataHelper.prepareFieldMappingDefectTypeTransformation(mapOfProjectFilters,
+						fieldMapping.getJiradefecttype(), fieldMapping.getJiraIssueTypeKPI171(),
+						JiraFeatureHistory.STORY_TYPE.getFieldValueInFeature());
+				uniqueProjectMap.put(basicProjectConfigId.toString(), mapOfProjectFilters);
+			}
+			List<String> status = new ArrayList<>();
+			if (Optional.ofNullable(fieldMapping.getJiraDodKPI171()).isPresent()) {
+				status.addAll(fieldMapping.getJiraDodKPI171());
+			}
 
-			KpiDataHelper.prepareFieldMappingDefectTypeTransformation(mapOfProjectFilters, fieldMapping.getJiradefecttype(),
-					fieldMapping.getJiraIssueTypeKPI171(), JiraFeatureHistory.STORY_TYPE.getFieldValueInFeature());
+			if (Optional.ofNullable(fieldMapping.getJiraDorKPI171()).isPresent()) {
+				status.addAll(fieldMapping.getJiraDorKPI171());
+			}
+
+			if (Optional.ofNullable(fieldMapping.getJiraLiveStatusKPI171()).isPresent()) {
+				status.addAll(fieldMapping.getJiraLiveStatusKPI171());
+			}
+			mapOfProjectFilters.put("statusUpdationLog.story.changedTo", CommonUtils.convertToPatternList(status));
 			uniqueProjectMap.put(basicProjectConfigId.toString(), mapOfProjectFilters);
-		}
-		List<String> status = new ArrayList<>();
-		if (Optional.ofNullable(fieldMapping.getJiraDodKPI171()).isPresent()) {
-			status.addAll(fieldMapping.getJiraDodKPI171());
-		}
 
-		if (Optional.ofNullable(fieldMapping.getJiraDorKPI171()).isPresent()) {
-			status.addAll(fieldMapping.getJiraDorKPI171());
-		}
-
-		if (Optional.ofNullable(fieldMapping.getJiraLiveStatusKPI171()).isPresent()) {
-			status.addAll(fieldMapping.getJiraLiveStatusKPI171());
-		}
-		mapOfProjectFilters.put("statusUpdationLog.story.changedTo", CommonUtils.convertToPatternList(status));
-		uniqueProjectMap.put(basicProjectConfigId.toString(), mapOfProjectFilters);
-
-		mapOfFilters.put(JiraFeature.BASIC_PROJECT_CONFIG_ID.getFieldValueInFeature(),
-				basicProjectConfigIds.stream().distinct().collect(Collectors.toList()));
-
-		List<JiraIssueCustomHistory> filteredProjectHistory = BacklogKpiHelper.filterProjectHistories(
-				getJiraIssuesCustomHistoryFromBaseClass(), uniqueProjectMap, startDate, endDate);
-
-		resultListMap.put(STORY_HISTORY_DATA, filteredProjectHistory);
+			List<JiraIssueCustomHistory> jiraIssueCustomHistoryList = jiraIssueCustomHistoryRepository
+					.findByBasicProjectConfigIdIn(basicProjectConfigId.toString());
+			List<JiraIssueCustomHistory> filteredProjectHistory = BacklogKpiHelper
+					.filterProjectHistories(jiraIssueCustomHistoryList, uniqueProjectMap, startDate, endDate);
+			resultListMap.put(basicProjectConfigId.toString(), filteredProjectHistory);
+		});
 		return resultListMap;
 	}
 
@@ -171,26 +222,12 @@ public class CycleTimeServiceImpl extends JiraBacklogKPIService<Long, List<Objec
 		return KPICode.CYCLE_TIME.name();
 	}
 
-	private void kpiWithFilter(Map<String, List<JiraIssueCustomHistory>> projectWiseJiraIssue, Node leafNode,
-			KpiElement kpiElement, DataCount dataCount) {
-
-		List<JiraIssueCustomHistory> issueCustomHistoryList = projectWiseJiraIssue
-				.get(leafNode.getProjectFilter().getBasicProjectConfigId().toString());
-
-		FieldMapping fieldMapping = configHelperService.getFieldMappingMap()
-				.get(leafNode.getProjectFilter().getBasicProjectConfigId());
-		List<CycleTimeValidationData> cycleTimeList = new ArrayList<>();
-		getCycleTime(issueCustomHistoryList, fieldMapping, cycleTimeList, kpiElement, dataCount);
-
-		kpiElement.setModalHeads(KPIExcelColumn.CYCLE_TIME.getColumns());
-	}
-
-	public void getCycleTime(List<JiraIssueCustomHistory> jiraIssueCustomHistoriesList, FieldMapping fieldMapping,
-			List<CycleTimeValidationData> cycleTimeList, KpiElement kpiElement, DataCount trendValue) {
+	public Map<String, List<DataValue>> getCycleTimeDataCount(
+			List<JiraIssueCustomHistory> jiraIssueCustomHistoriesList, FieldMapping fieldMapping,
+			List<CycleTimeValidationData> cycleTimeList, KpiElement kpiElement) {
 
 		Set<String> issueTypeFilter = new LinkedHashSet<>();
-
-		List<IterationKpiValue> dataList = new ArrayList<>();
+		Map<String, List<DataValue>> cycleMap = new LinkedHashMap<>();
 
 		if (CollectionUtils.isNotEmpty(jiraIssueCustomHistoriesList)) {
 			List<Long> overAllIntakeDorTime = new ArrayList<>();
@@ -248,9 +285,12 @@ public class CycleTimeServiceImpl extends JiraBacklogKPIService<Long, List<Objec
 
 						jiraIssueCustomHistory.getStatusUpdationLog().forEach(statusUpdateLog -> {
 							DateTime updateTime = DateTime.parse(statusUpdateLog.getUpdatedOn().toString());
-							BacklogKpiHelper.setLiveTime(cycleTimeValidationData, cycleTime, statusUpdateLog, updateTime, liveStatus);
-							BacklogKpiHelper.setReadyTime(cycleTimeValidationData, cycleTime, statusUpdateLog, updateTime, dor);
-							BacklogKpiHelper.setDODTime(statusUpdateLog, updateTime, dodStatus, storyFirstStatus, dodStatusDateMap);
+							BacklogKpiHelper.setLiveTime(cycleTimeValidationData, cycleTime, statusUpdateLog,
+									updateTime, liveStatus);
+							BacklogKpiHelper.setReadyTime(cycleTimeValidationData, cycleTime, statusUpdateLog,
+									updateTime, dor);
+							BacklogKpiHelper.setDODTime(statusUpdateLog, updateTime, dodStatus, storyFirstStatus,
+									dodStatusDateMap);
 						});
 						DateTime minUpdatedOn = CollectionUtils.isNotEmpty(dodStatusDateMap.values())
 								? Collections.min(dodStatusDateMap.values())
@@ -265,80 +305,82 @@ public class CycleTimeServiceImpl extends JiraBacklogKPIService<Long, List<Objec
 								cycleTime.getDeliveryTime(), DOR_TO_DOD_KPI, cycleTimeValidationData, null);
 						String deliverToLive = BacklogKpiHelper.setValueInCycleTime(cycleTime.getDeliveryTime(),
 								cycleTime.getLiveTime(), DOD_TO_LIVE_KPI, cycleTimeValidationData, null);
-						String leadTime = BacklogKpiHelper.setValueInCycleTime(cycleTime.getIntakeTime(), cycleTime.getLiveTime(),
-								LEAD_TIME_KPI, cycleTimeValidationData, null);
-						transitionExist(overAllIntakeDorTime, overAllIntakeDorModalValues, intakeDorTime, intakeDorModalValues,
-								jiraIssueCustomHistory, intakeToReady);
+						String leadTime = BacklogKpiHelper.setValueInCycleTime(cycleTime.getIntakeTime(),
+								cycleTime.getLiveTime(), LEAD_TIME_KPI, cycleTimeValidationData, null);
+						transitionExist(overAllIntakeDorTime, overAllIntakeDorModalValues, intakeDorTime,
+								intakeDorModalValues, jiraIssueCustomHistory, intakeToReady);
 						transitionExist(overAllDorDodTime, overAllDorDodModalValues, dorDodTime, dorDodModalValues,
 								jiraIssueCustomHistory, readyToDeliver);
 						transitionExist(overAllDodLiveTime, overAllDodLiveModalValues, dodLiveTime, dodLiveModalValues,
 								jiraIssueCustomHistory, deliverToLive);
-						transitionExist(overAllLeadTimeList, overAllLeadTimeModalValues, leadTimeList, leadTimeModalValues,
-								jiraIssueCustomHistory, leadTime);
+						transitionExist(overAllLeadTimeList, overAllLeadTimeModalValues, leadTimeList,
+								leadTimeModalValues, jiraIssueCustomHistory, leadTime);
 
 						cycleTimeList.add(cycleTimeValidationData);
 					}
 
 					intakeDor = AggregationUtils.averageLong(intakeDorTime);
+					cycleMap.put(INTAKE_TO_DOR + "#" + type, Arrays.asList(getDataValue(intakeDor, "Issues"),
+							getDataValue(ObjectUtils.defaultIfNull(intakeDor, 0L) / 480, "Days")));
 					dorDod = AggregationUtils.averageLong(dorDodTime);
+					cycleMap.put(DOR_TO_DOD + "#" + type, Arrays.asList(getDataValue(dorDod, "Issues"),
+							getDataValue(ObjectUtils.defaultIfNull(dorDod, 0L) / 480, "Days")));
 					dodLive = AggregationUtils.averageLong(dodLiveTime);
+					cycleMap.put(DOD_TO_LIVE + "#" + type, Arrays.asList(getDataValue(dodLive, "Issues"),
+							getDataValue(ObjectUtils.defaultIfNull(dodLive, 0L) / 480, "Days")));
 
-					List<IterationKpiData> data = new ArrayList<>();
-					data.add(new IterationKpiData(INTAKE_TO_DOR, ObjectUtils.defaultIfNull(intakeDor, 0L).doubleValue() / 480,
-							ObjectUtils.defaultIfNull(intakeDorTime.size(), 0L).doubleValue(), null, DAYS, ISSUES,
-							getIterationKpiModalValue(intakeDorModalValues, cycleTimeList)));
-					data.add(new IterationKpiData(DOR_TO_DOD, ObjectUtils.defaultIfNull(dorDod, 0L).doubleValue() / 480,
-							ObjectUtils.defaultIfNull(dorDodTime.size(), 0L).doubleValue(), null, DAYS, ISSUES,
-							getIterationKpiModalValue(dorDodModalValues, cycleTimeList)));
-					data.add(new IterationKpiData(DOD_TO_LIVE, ObjectUtils.defaultIfNull(dodLive, 0L).doubleValue() / 480,
-							ObjectUtils.defaultIfNull(dodLiveTime.size(), 0L).doubleValue(), null, DAYS, ISSUES,
-							getIterationKpiModalValue(dodLiveModalValues, cycleTimeList)));
-					IterationKpiValue iterationKpiValue = new IterationKpiValue(type, null, data);
-					dataList.add(iterationKpiValue);
 				}
 			});
 			overAllIntakeDor = AggregationUtils.averageLong(overAllIntakeDorTime);
+			cycleMap.put(INTAKE_TO_DOR + "#Overall", Arrays.asList(getDataValue(overAllIntakeDor, "Issues"),
+					getDataValue(ObjectUtils.defaultIfNull(overAllIntakeDor, 0L) / 480, "Days")));
 			overAllDorDod = AggregationUtils.averageLong(overAllDorDodTime);
+			cycleMap.put(DOR_TO_DOD + "#Overall", Arrays.asList(getDataValue(overAllDorDod, "Issues"),
+					getDataValue(ObjectUtils.defaultIfNull(overAllDorDod, 0L) / 480, "Days")));
 			overAllDodLive = AggregationUtils.averageLong(overAllDodLiveTime);
-
-			List<IterationKpiData> data = new ArrayList<>();
-			data.add(new IterationKpiData(INTAKE_TO_DOR, ObjectUtils.defaultIfNull(overAllIntakeDor, 0L).doubleValue() / 480,
-					ObjectUtils.defaultIfNull(overAllIntakeDorTime.size(), 0L).doubleValue(), null, DAYS, ISSUES,
-					getIterationKpiModalValue(overAllIntakeDorModalValues, cycleTimeList)));
-			data.add(new IterationKpiData(DOR_TO_DOD, ObjectUtils.defaultIfNull(overAllDorDod, 0L).doubleValue() / 480,
-					ObjectUtils.defaultIfNull(overAllDorDodTime.size(), 0L).doubleValue(), null, DAYS, ISSUES,
-					getIterationKpiModalValue(overAllDorDodModalValues, cycleTimeList)));
-			data.add(new IterationKpiData(DOD_TO_LIVE, ObjectUtils.defaultIfNull(overAllDodLive, 0L).doubleValue() / 480,
-					ObjectUtils.defaultIfNull(overAllDodLiveTime.size(), 0L).doubleValue(), null, DAYS, ISSUES,
-					getIterationKpiModalValue(overAllDodLiveModalValues, cycleTimeList)));
-			IterationKpiValue iterationKpiValue = new IterationKpiValue(CommonConstant.OVERALL, CommonConstant.OVERALL, data);
-			dataList.add(iterationKpiValue);
+			cycleMap.put(DOD_TO_LIVE + "#Overall", Arrays.asList(getDataValue(overAllDodLive, "Issues"),
+					getDataValue(ObjectUtils.defaultIfNull(overAllDodLive, 0L) / 480, "Days")));
 		}
-		Set<String> duration = new LinkedHashSet<>(
-				Arrays.asList("Past Week", "Past 2 Weeks", "Past Month", "Past 3 Months", "Past 6 Months"));
-		IterationKpiFiltersOptions filter1 = new IterationKpiFiltersOptions(SEARCH_BY_DURATION, duration);
+		IterationKpiFiltersOptions filter1 = new IterationKpiFiltersOptions(SEARCH_BY_DURATION, durationFilter);
 		IterationKpiFiltersOptions filter2 = new IterationKpiFiltersOptions(SEARCH_BY_ISSUE_TYPE, issueTypeFilter);
 		IterationKpiFilters iterationKpiFilters = new IterationKpiFilters(filter1, filter2);
 		// Modal Heads Options
 		kpiElement.setFilters(iterationKpiFilters);
-		trendValue.setValue(dataList);
-		kpiElement.setTrendValueList(trendValue);
+
+		return cycleMap;
+
 	}
 
-	private List<IterationKpiModalValue> getIterationKpiModalValue(List<JiraIssueCustomHistory> modalJiraIssues,
-			List<CycleTimeValidationData> cycleTimeList) {
-		List<IterationKpiModalValue> iterationKpiDataList = new ArrayList<>();
-		Map<String, IterationKpiModalValue> dataMap = KpiDataHelper.createMapOfModalObjectFromJiraHistory(modalJiraIssues,
-				cycleTimeList);
-		for (Map.Entry<String, IterationKpiModalValue> entry : dataMap.entrySet()) {
-			iterationKpiDataList.add(dataMap.get(entry.getKey()));
+	private Map<String, List<DataCount>> getDataCountObject(String trendLineName, Map<String, List<DataValue>> cycleMap, String duration) {
+		Map<String, List<DataCount>> dataCountMap = new HashMap<>();
+		cycleMap.forEach((key, value) -> {
+			String[] issueFilter = key.split("#");
+			String filterKey = duration+"#"+issueFilter[1];
+			DataCount dataCount = new DataCount();
+			dataCount.setSProjectName(trendLineName);
+			dataCount.setDataValue(value);
+			dataCount.setKpiGroup(filterKey);
+			dataCount.setSubFilter(issueFilter[0]);
+			dataCount.setHoverValue(new HashMap<>());
+			dataCountMap.computeIfAbsent(filterKey, k -> new ArrayList<>()).add(dataCount);
+		});
+		return dataCountMap;
+	}
+
+	private DataValue getDataValue(Long overAllCount, String subFilter) {
+		DataValue dataValue = new DataValue();
+		dataValue.setData(String.valueOf(overAllCount));
+		dataValue.setValue(overAllCount);
+		dataValue.setName(subFilter);
+		return dataValue;
+
+	}
+
+	private void populateExcelDataObject(String requestTrackerId, List<CycleTimeValidationData> cycleTimeList,
+			List<KPIExcelData> excelData) {
+		if (requestTrackerId.toLowerCase().contains(KPISource.EXCEL.name().toLowerCase())) {
+			KPIExcelUtility.populateCycleTime(cycleTimeList, excelData);
 		}
-		return iterationKpiDataList;
-	}
-
-	@Override
-	public Long calculateKpiValue(List<Long> valueList, String kpiName) {
-		return calculateKpiValueForLong(valueList, kpiName);
 	}
 
 	private void transitionExist(List<Long> overAllTimeList, List<JiraIssueCustomHistory> overAllTransitionModalValues,
